02_intermediate_agent/api/requirements.txt
fastapi[standard]>=0.110.0
langchain-core>=0.1.30
langgraph>=0.2.0
litellm>=1.34.0
pydantic>=2.6.0
python-dotenv>=1.0.1
uvicorn>=0.28.0
